requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.7
django-rest-authtoken==2.1.4
djangorestframework==3.12.4
idna==3.2
psycopg2==2.9.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
twilio==7.0.0
urllib3==1.26.7
